+++ lib/ui/signup/profile/sign_up_profile_screen.dart
@@ -1,0 +1,15 @@
+
+
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class SignUpProfileScreen extends StatelessWidget{
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+
+    );
+  }
+
+}
